+++ src/utils/menu.ts
@@ -1,5 +1,6 @@
-import { nextTick, reactive, ref } from 'vue'
+import { nextTick, reactive, computed } from 'vue'
 import { useRouter, useRoute } from 'vue-router'
+import { useStore } from 'vuex'
 import config from '@/utils/menu-config'
 
 const WIDTH = 1200 // 菜单消失和出现阀值
@@ -7,6 +8,7 @@ const { body } = document
 export default () => {
   const router = useRouter()
   const route = useRoute()
+  const store = useStore()
 
   const menuConfig = reactive(config)
 
@@ -74,20 +76,30 @@ export default () => {
     }
   }
 
-  const isMenuRective = ref(false) // 标记当前菜单栏是否可显示隐藏的
+  const isMenuRective = computed(() => store.state.menuStore.isMenuRective) // 标记当前菜单栏是否可显示隐藏的
+  const isFolded = computed(() => store.state.menuStore.isFolded) // 菜单栏隐藏
   /**
    * 判断当前菜单栏是否需要显示隐藏
    */
   const setMenuRective = () => {
     const rect = body.getBoundingClientRect()
-    isMenuRective.value = rect.width <= WIDTH
+    store.commit('setMenuRective', rect.width <= WIDTH)
+    store.commit('setFolded', true)
+  }
+  /**
+   * 切换菜单栏的隐藏显示
+   */
+  const toggleFold = () => {
+    store.commit('setFolded', !isFolded.value)
   }
   return {
     menuConfig,
     isMenuRective,
+    isFolded,
     getCurrMenu,
     clickMenu,
     changePath,
-    setMenuRective
+    setMenuRective,
+    toggleFold
   }
 }
